requirements.txt
google-genai
neo4j
streamlit
pandas
python-dotenv
networkx
streamlit-agraph
